app/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
pysnmp>=6.2
apscheduler==3.10.4
pyyaml==6.0.2
python-dotenv==1.0.1
gunicorn==23.0.0
dnspython==2.7.0
